resources/cdk/rekognition-photo-analyzer/requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
types-boto3
types-setuptools
black